feat: validar existencia del archivo antes de procesarlo

+++ modules/Sale/Services/SaleOpportunityFileService.php
@@ -9,7 +9,13 @@ class SaleOpportunityFileService
 
     public function getFile($filename)
     {
-        $file =  Storage::disk('tenant')->get('sale_opportunity_files'.DIRECTORY_SEPARATOR.$filename);
+        $path = 'sale_opportunity_files' . DIRECTORY_SEPARATOR . $filename;
+
+        if (!Storage::disk('tenant')->exists($path)) {
+            die('El archivo no existe.');
+        }
+    
+        $file = Storage::disk('tenant')->get($path);
         $temp = tempnam(sys_get_temp_dir(), 'tmp_sale_opportunity_files');
         file_put_contents($temp, $file);
         $mime = mime_content_type($temp);
